+++ offsets-tracker/cache/cache.go
@@ -1,3 +1,17 @@
+// Copyright The OpenTelemetry Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package cache
 
 import (
@@ -11,10 +25,12 @@ import (
 	"go.opentelemetry.io/auto/offsets-tracker/schema"
 )
 
+// Cache holds already seen offsets.
 type Cache struct {
 	data *schema.TrackedOffsets
 }
 
+// NewCache returns a new [Cache].
 func NewCache(prevOffsetFile string) *Cache {
 	f, err := os.Open(prevOffsetFile)
 	if err != nil {
@@ -41,6 +57,8 @@ func NewCache(prevOffsetFile string) *Cache {
 	}
 }
 
+// IsAllInCache returns all DataMemberOffset for a module with modName and
+// version and offsets describe by dm.
 func (c *Cache) IsAllInCache(modName string, version string, dm []*binary.DataMember) ([]*binary.DataMemberOffset, bool) {
 	for _, item := range c.data.Data {
 		if item.Name == modName {
@@ -63,9 +81,8 @@ func (c *Cache) IsAllInCache(modName string, version string, dm []*binary.DataMe
 			}
 			if offsetsFound == len(dm) {
 				return results, true
-			} else {
-				return nil, false
 			}
+			return nil, false
 		}
 	}