+++ offsets-tracker/writer/writer.go
@@ -1,3 +1,17 @@
+// Copyright The OpenTelemetry Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package writer
 
 import (
@@ -13,6 +27,7 @@ import (
 	"go.opentelemetry.io/auto/offsets-tracker/target"
 )
 
+// WriteResults writes results to fileName.
 func WriteResults(fileName string, results ...*target.Result) error {
 	var offsets schema.TrackedOffsets
 	for _, r := range results {
@@ -42,13 +57,13 @@ func WriteResults(fileName string, results ...*target.Result) error {
 		return err
 	}
 
-	var prettyJson bytes.Buffer
-	err = json.Indent(&prettyJson, jsonData, "", "  ")
+	var prettyJSON bytes.Buffer
+	err = json.Indent(&prettyJSON, jsonData, "", "  ")
 	if err != nil {
 		return err
 	}
 
-	return os.WriteFile(fileName, prettyJson.Bytes(), fs.ModePerm)
+	return os.WriteFile(fileName, prettyJSON.Bytes(), fs.ModePerm)
 }
 
 func convertResult(r *target.Result) schema.TrackedLibrary {