+++ pkg/instrumentors/allocator/allocator_linux.go
@@ -17,18 +17,22 @@ package allocator
 import (
 	"os"
 
+	"golang.org/x/sys/unix"
+
 	"go.opentelemetry.io/auto/pkg/instrumentors/bpffs"
 	"go.opentelemetry.io/auto/pkg/instrumentors/context"
 	"go.opentelemetry.io/auto/pkg/log"
-	"golang.org/x/sys/unix"
 )
 
+// Allocator handles the allocation of the BPF file-system.
 type Allocator struct{}
 
+// New returns a new [Allocator].
 func New() *Allocator {
 	return &Allocator{}
 }
 
+// Load loads the BPF file-system.
 func (a *Allocator) Load(ctx *context.InstrumentorContext) error {
 	logger := log.Logger.WithName("allocator")
 	if ctx.TargetDetails.AllocationDetails != nil {
@@ -38,7 +42,7 @@ func (a *Allocator) Load(ctx *context.InstrumentorContext) error {
 	}
 	logger.V(0).Info("Loading allocator")
 
-	err := a.mountBpfFS()
+	err := a.mountBPFFs()
 	if err != nil {
 		return err
 	}
@@ -46,17 +50,16 @@ func (a *Allocator) Load(ctx *context.InstrumentorContext) error {
 	return nil
 }
 
-func (a *Allocator) mountBpfFS() error {
-	_, err := os.Stat(bpffs.BpfFsPath)
+func (a *Allocator) mountBPFFs() error {
+	_, err := os.Stat(bpffs.BPFFsPath)
 	if err != nil {
-		if os.IsNotExist(err) {
-			if err := os.MkdirAll(bpffs.BpfFsPath, 0755); err != nil {
-				return err
-			}
-		} else {
+		if !os.IsNotExist(err) {
+			return err
+		}
+		if err := os.MkdirAll(bpffs.BPFFsPath, 0755); err != nil {
 			return err
 		}
 	}
 
-	return unix.Mount(bpffs.BpfFsPath, bpffs.BpfFsPath, "bpf", 0, "")
+	return unix.Mount(bpffs.BPFFsPath, bpffs.BPFFsPath, "bpf", 0, "")
 }