+++ pkg/instrumentors/bpf/google/golang/org/grpc/probe.go
@@ -43,8 +43,8 @@ type GrpcEvent struct {
 	EndTime           uint64
 	Method            [50]byte
 	Target            [50]byte
-	SpanContext       context.EbpfSpanContext
-	ParentSpanContext context.EbpfSpanContext
+	SpanContext       context.EBPFSpanContext
+	ParentSpanContext context.EBPFSpanContext
 }
 
 type grpcInstrumentor struct {
@@ -88,7 +88,7 @@ func (g *grpcInstrumentor) Load(ctx *context.InstrumentorContext) error {
 	g.bpfObjects = &bpfObjects{}
 	err = spec.LoadAndAssign(g.bpfObjects, &ebpf.CollectionOptions{
 		Maps: ebpf.MapOptions{
-			PinPath: bpffs.BpfFsPath,
+			PinPath: bpffs.BPFFsPath,
 		},
 	})
 	if err != nil {